+++ class-components/src/components/index.ts
@@ -1,11 +1,11 @@
-import { SearchBar } from './SearchBar.tsx';
-import { List } from './List.tsx';
-import { ListItem } from './ListItem.tsx';
-import { Loader } from './Loader.tsx';
+import { SearchBar } from './SearchBar/SearchBar.tsx';
+import { List } from './List/List.tsx';
+import { ListItem } from './ListItem/ListItem.tsx';
+import { Loader } from './Loader/Loader.tsx';
 import { ErrorBoundary } from './ErrorBoundary.tsx';
 import { AppRouter } from './AppRouter.tsx';
-import { ListItemDetailed } from './ListItemDetailed.tsx';
-import { Pagination } from './Pagination.tsx';
-import { Header } from './Header.tsx';
+import { ListItemDetailed } from './ListItemDetailed/ListItemDetailed.tsx';
+import { Pagination } from './Pagination/Pagination.tsx';
+import { Header } from './Header/Header.tsx';
 
 export { SearchBar, List, ListItem, Loader, ErrorBoundary, AppRouter, ListItemDetailed, Pagination, Header };
